scripts/scraper/requirements.txt
supabase
trafilatura
beautifulsoup4
python-dotenv
requests
lxml_html_clean
